Live performance modulation cuts current segment short if possible (#95)
In manual-control mode, when the user engages the next set of parameters, we will switch to it as quickly as possible-- this will rely on the subdivisions to only change at the top of a bar (e.g. at 4 beats per bar in a 16 beat sequence, there are 4 bars).

- Determine the segment we are currently in the middle of dubbing
- Determine whether the current segment can be cut short
- Cut the current segment short after the given number of beats
- Delete all segments after the current segment and fabricate the next segment
- Re-render the new segment that's been cut short (should show new shorter total number of bars)

https://www.pivotaltracker.com/story/show/186686015

+++ gui/src/test/java/io/xj/gui/services/FabricationServiceImplTest.java
@@ -46,7 +46,7 @@ class FabricationServiceImplTest {
   int defaultOutputChannels = 2;
   private final String defaultMacroMode = ControlMode.AUTO.toString();
   String defaultInputMode = InputMode.PRODUCTION.toString();
-  double defaultOutputFrameRate = 48000;
+  int defaultOutputFrameRate = 48000;
 
   @Mock
   HostServices hostServices;
